feat: optional verbose logging, highlight file with errors in reports.
if the VERBOSE env var is set, the command will produce more verbose
logging, but this log can help to catch files that needs to be excluded
from the coverage collection (e.g. build files, minified builds), which
most of the time flood the flow deamon and the coverage collection never
completes.