No annotation, no coverage
When calculating coverage for a project, I don't care about coverage in
files that don't have the `@flow` pragma. While it's nice that Flow
infers types anyway, since it's never going to report an error it
doesn't help me at all and is potentially misleading. For example, my
project is 75% covered with master, but only 34% covered with this
change.

This PR changes flow-coverage-report to ignore coverage in unannotated
files. This could be considered a breaking change or a bugfix - I'm not
sure. You might consider making this a configuration option.

Resolves #85